refactor: better code management for AgentCapabilities and ToolGroups
