+++ refact-agent/gui/src/components/ChatForm/AgentCapabilities/index.ts
@@ -1,0 +1,1 @@
+export { AgentCapabilities } from "./AgentCapabilities";
